mock_daos.go moved into internal/testutils/mocks
